requirements.txt
langchain
psycopg2
python-dotenv